Merged PR 34660: Updated UI test and added 1 API test for MP
- UI Tests update
Previously I was searching by a stockcode and then adding it to the cart. Tests would fail when the product was unavailable. Now, I search for a keyword and add any available product. The test is dynamic and will adapt to the available product and its current price as well as any promotion on it.

- Added a new API test for MP