Merged PR 42406: Fix for cancel order
The test cases for cancel order were failing because searchBody.SearchTerm=Kitchen in createorder.js, and currently, while adding the items to the trolley , the subtotal was equal to $27. Resulting it failed as it was expecting more or equal to $30.
As a fix/workaround, have changed the value for searchBody.SearchTerm