Merged PR 50989: Verify DU discount In Progress
This PR is to provide 2 generic methods to team to add MultiSeller EM products to the Trolley, which is added as part of DU discount script which is still in progress.
1- prepareAnyMultiSellerLineItemWowAndEdmOrder() adds the Wow + EM Multi Seller Items in the Cart
2- prepareAnyMultiSellerLineItemEdmOrder() Add EM ONLY Multi Seller Items in the Cart //currently commented in this PR
3- Using 'emAccountWithRewards27' present in shoppers.json file
The script just places the Order currently. DU functionality will be built in future PRs on top of this.